Add a basic TrajectoryGenerator module
With this we can find the current curve of the segment we are driving
on. Bezier curves can be imported from file and stored in a new segment
class. The location of the file and the segment to drive on can be
specified in the config file. For this a new "eagle_eye.cfg" config
file has been created.

(sg, fh)