Improve logtosvg with polyline elements
Instead with single circles the data points from the log files are now
visualized with one polyline for every file. This results in a much
smaller svg file.

Dealing with units and scaling is improved, so the data fits on a smaller
canvas and can be better displayed, without worrying about units.