Test shifted coordinate system on car (fh)
This provides a modification where the whole position of the car is
shifted between the back wheels. Controller parameters are choosen by
Matlab LQR (r=1e-10). Works great in simulation.

Also, state.position and projected point of the bezier curve are now
shown in the simulation. (position=black, projected=green)