Clean up U/ directory, add TrafficLight Modul (fh)
Various files are set back to an previous state and some are added
to .gitignore.

Ampelerkennung.cpp/h, RedCircleDetection and StopAtRed are moved out
of the project file and out of directory der_weg/ and are now found in
U/Ampelerkennung_test/.

A new Modul TrafficLightDetection is created, which is intended to
hold all functionality for detecting the traffic lights.