Add first traffic light localization (fh)
In TrafficLightDetection there is now a localization, which estimates
the traffic light positions by iterative averaging over the
measruements. The covariances are incoporated in the calculations.

The initial estimated positions, along with the initial covariance can
be specified in the config. Also there: the covariance of the
measurements.

The currently measure position together with the estimated ones is
displayed in the AnicarViewer as dots.
(measurement = red, estmates = dark green)