Further implementation of traffic light detection
Distances to detected ellipses is now calculated via the median of the
pixels of the rectangular bounding box applied to the depth image.

The matrix for transforming image coordinates to camera coordinates is
now available. It is used to calculate the camera coordinates of the
ellipse center points.

Fix the erode_size bug.

(jf, fh)