Add checks for bounding box and confidence (fh)
The ROI for extracting depth information is now limited to the right and
left boarders of the image. The upper and lower boarder are not
considered.

For the remaining ROI the conidence must exceed a given minimal value
(config). Pixels with lower confidence are not considered.
If no Pixel remains, the distance is instead calculated via the heigth
of the ROI in the image.
This method is still to be implemented.