Add module for image rectification (fh)
Rectifier is a new module which does the rectification of the left and
right images and writes them to the blackboard as an RectImages data
object. TLD and ConeDetection then read this object.

Together with the rectified images the state and the reference
trajectory are provided to ensure they all refer to the same point in
time.